larger datatype for Cost

+++ include/mad_types.h
@@ -14,6 +14,6 @@ typedef int Level;
 typedef int ProcessID;
 
 /// Cost (compute cost) of a node (can be changed but set to integer for now)
-typedef int Cost;
+typedef unsigned long Cost;
 
 #endif
